Refactor workflow data sanitization to use a shared utility class
