+++ src/Modules/Workflows/Rest/RestApiManager.php
@@ -5,6 +5,7 @@ namespace PublishPress\Future\Modules\Workflows\Rest;
 use PublishPress\Future\Core\HookableInterface;
 use PublishPress\Future\Modules\Workflows\HooksAbstract;
 use PublishPress\Future\Modules\Workflows\Interfaces\RestApiManagerInterface;
+use PublishPress\Future\Framework\WordPress\Utils\WorkflowSanitizationUtil;
 
 // TODO: This is a temporary class to register the REST API routes. Move this to controllers on each module.
 class RestApiManager implements RestApiManagerInterface
@@ -16,15 +17,23 @@ class RestApiManager implements RestApiManagerInterface
      */
     private HookableInterface $hooks;
 
-    public function __construct(HookableInterface $hooks)
-    {
+    private WorkflowSanitizationUtil $workflowSanitization;
+
+    public function __construct(
+        HookableInterface $hooks,
+        WorkflowSanitizationUtil $workflowSanitization
+    ) {
         $this->hooks = $hooks;
+        $this->workflowSanitization = $workflowSanitization;
     }
 
     public function register()
     {
         $apiManagers = [
-            new RestApiV1($this->hooks)
+            new RestApiV1(
+                $this->hooks,
+                $this->workflowSanitization
+            )
         ];
 
         /**
